+++ node/src/main/java/com/topchain/node/model/viewModel/BlockViewModel.java
@@ -8,7 +8,7 @@ import java.util.Set;
 public class BlockViewModel {
     private Long index;
     private Set<Transaction> transactions;
-    private Long difficulty;
+    private int difficulty;
     private String previousBlockHash;
     private String minedBy;
     private String blockDataHash;
@@ -35,11 +35,11 @@ public class BlockViewModel {
         this.transactions = transactions;
     }
 
-    public Long getDifficulty() {
+    public int getDifficulty() {
         return difficulty;
     }
 
-    public void setDifficulty(Long difficulty) {
+    public void setDifficulty(int difficulty) {
         this.difficulty = difficulty;
     }
